+++ TripLog/TripLog.xcodeproj/project.xcworkspace/xcshareddata/swiftpm/Package.resolved
@@ -1,5 +1,5 @@
 {
-  "originHash" : "d5fc5c9f1d3bc526f11564c096e47cb79140e9216b7b53f033f45ed8c25e26d1",
+  "originHash" : "8e3df952529c2a86bc250ddeb8f702b5a78e13ed260b48ce6316099ee638e17f",
   "pins" : [
     {
       "identity" : "alamofire",
