+++ TripLog/TripLog/Feat-Stone/View/TopViewController.swift
@@ -15,12 +15,11 @@ class TopViewController: UIViewController, UITableViewDataSource, UITableViewDel
         $0.separatorStyle = .none
     }
 
-    // Model 데이터
     private let data = TestDummyData.sampleData() // Model에서 가져옴
 
     override func viewDidLoad() {
         super.viewDidLoad()
-        view.backgroundColor = .white
+        view.applyBackgroundColor()
 
         navigationController?.navigationBar.isHidden = false
 
@@ -29,28 +28,25 @@ class TopViewController: UIViewController, UITableViewDataSource, UITableViewDel
             .font: UIFont.SCDream(size: .title, weight: .bold)
         ]
 
-        // 초기 네비게이션 바 제목 설정 (첫 번째 여행 제목)
         if let firstTrip = data.first {
             self.navigationItem.title = firstTrip.title
         }
+        
+        tableView.rowHeight = UITableView.automaticDimension
+        tableView.estimatedRowHeight = UIScreen.main.bounds.height * 0.5
 
         setupTableView()
     }
 
     private func setupTableView() {
-        // 테이블 뷰 추가
         view.addSubview(tableView)
 
-        // 테이블 뷰 설정
         tableView.dataSource = self
         tableView.delegate = self
         tableView.register(CustomTableViewCell.self, forCellReuseIdentifier: "CustomCell")
 
-        // SnapKit 레이아웃 설정
         tableView.snp.makeConstraints {
-            $0.top.equalToSuperview().offset(20)
-            $0.leading.trailing.equalToSuperview()
-            $0.bottom.equalToSuperview().offset(-20)
+            $0.edges.equalToSuperview()
         }
     }
 
@@ -75,54 +71,12 @@ class TopViewController: UIViewController, UITableViewDataSource, UITableViewDel
     // MARK: - UITableViewDelegate
 
     func tableView(_ tableView: UITableView, didSelectRowAt indexPath: IndexPath) {
-        // 셀을 선택했을 때
         tableView.deselectRow(at: indexPath, animated: true)
 
-        // 선택된 trip 데이터 가져오기
         let selectedTrip = data[indexPath.row]
-
-        // 네비게이션 바의 제목을 선택된 trip의 제목으로 설정
         self.navigationItem.title = selectedTrip.title
 
         print("Cell tapped: \(indexPath.row), Title: \(selectedTrip.title)")
-    }
-}
-
-class CustomTableViewCell: UITableViewCell {
-
-    private let titleDateView = TitleDateView()
-    private let progressView = ProgressView()
-    private let buttonStackView = CustomButtonStackView()
-
-    func configure(subtitle: String, date: String, expense: String, budget: String) {
-        // 데이터 설정
-        titleDateView.configure(subtitle: subtitle, date: date)
-        progressView.configure(expense: expense, budget: budget)
-
-        setupLayout()
-    }
-
-    private func setupLayout() {
-        // 모든 서브뷰 추가
-        [titleDateView, progressView, buttonStackView].forEach {
-            contentView.addSubview($0)
-        }
-
-        titleDateView.snp.makeConstraints {
-            $0.top.equalToSuperview().offset(4)
-            $0.leading.trailing.equalToSuperview().inset(16)
-        }
-
-        progressView.snp.makeConstraints {
-            $0.top.equalTo(titleDateView.snp.bottom).offset(16)
-            $0.leading.trailing.equalToSuperview().inset(16)
-        }
-
-        buttonStackView.snp.makeConstraints {
-            $0.top.equalTo(progressView.snp.bottom).offset(16)
-            $0.leading.trailing.equalToSuperview()
-            $0.height.equalTo(50)
-        }
     }
 }
